Add README and furo theme for docs (#123)
Play around with furo theme, resolves #118 

Note that the class sections aren't quite as distinct now in autoapi
docs, see below for an example:

Before

![image](https://github.com/user-attachments/assets/d4ebacb4-6395-4bcf-a0a0-97de7672afa3)

After

![image](https://github.com/user-attachments/assets/f3a7d2fa-81d0-4158-9458-bd20691ccf61)


Also get README into docs. This is a bit of a hack using a symlink,
generally, we don't have a good way to get docs outside of the `docs/`
directory into our docs. We could consider using the entire repo for
this?