Prepare rules_uv `create_venv` and add `pip` to requirements.in (#200)
See #202 for the full decision and distinction of the various options we
have for configuring a venv. tl;dr: we'll use `rules_uv` for a repo-wide
venv. To fully take advantage of it, using generated files, we'll pip
install a generated package for the repo there. This will enable us to
modify un-generated source and have it reflected in the venv's view of
things, while also getting generated code in there, such as protobufs.