simpy example (#225)
Follow tutorial for simpy
https://simpy.readthedocs.io/en/latest/simpy_intro/index.html and then
experiment with my own thermostat.

This PR is mostly meant to capture examples for how to use simpy.

While playing with this, I was struck by how one could structure an
arbitrary service in a way that could be easily run in a Discrete Event
Simulation (DES) context. Would involve wrapping/adapting all external
calls (network, time, system, thread prioritization, etc) and feeding
into a centralized broker. I imagine that it would have difficulty
dealing with "new processes", ones that were injected after the system
had begun to run, but it doesn't seem impossible, you'd just have to get
any new process also on this event loop.

To deal with the overall process execution, I wonder if we could
leverage
https://developers.facebook.com/blog/post/2022/11/22/hermit-deterministic-linux-testing/
or if it'd all have to be located in a centralized scheduler.

This seems to echo some of the principles inherent in event programming,
ala
https://berb.github.io/diploma-thesis/original/043_threadsevents.html