Update it all (#233)
![hal_changing_lightbulb](https://github.com/user-attachments/assets/029fa0c6-9489-4088-b278-317074e2f8aa)

I wanted to checkout `gazelle_cc`, but started having issues with
various bazel modules and the current bazel version I'm on. After
attempting a few hacks, I endeavoured to update to Bazel 8 and update
most of my packages to their "latest"ish versions, in the hopes that
this would lead to fewer cross-couplings.

Ironically, while doing this, I substituted gazelle's binary out, due to
the go compilation of tree sitter failing. With this substitution, I
can't configure the languages. So, I'll have to revisit.

#### To Do
- [x] Get tests passing
- [x] Make a table of third party dependencies, describe relationships,
and provide links
- [x] Clean up the changes and reorganize, etc.
- [x] check gazelle, etc. still functions